refactor(openapi): move openapi_routes to mockforge-openapi — Phase 4c
Physically relocates \`openapi_routes.rs\` + \`openapi_routes/\` (the
3887-line router file + its four submodules: \`builder\`, \`generation\`,
\`registry\`, \`validation\`) from mockforge-core into mockforge-openapi,
completing the original Phase 4 goal from the extraction plan.

The move was unblocked by Phase 4b, which swapped the handler's direct
calls to core's \`templating::expand_tokens\` and \`Overrides::apply\` for
dispatch through the \`ResponseRewriter\` trait. With that done, every
\`crate::\` path in \`openapi_routes\` now resolves either inside
mockforge-openapi itself or via foundation:

Rewritten imports (core -> foundation / openapi-internal):
- \`crate::ai_response::\` -> \`mockforge_foundation::ai_response::\`
- \`crate::reality_continuum::response_trace::\` ->
  \`mockforge_foundation::response_generation_trace::\`
- \`crate::schema_diff::\` -> \`mockforge_foundation::schema_diff::\`
- \`crate::latency::\` -> \`mockforge_foundation::latency::\`
- \`crate::FailureInjector\` ->
  \`mockforge_foundation::failure_injection::FailureInjector\`
- \`crate::CustomFixtureLoader\` ->
  \`crate::custom_fixture::CustomFixtureLoader\` (now in openapi)
- \`crate::RequestFingerprint\` ->
  \`crate::request_fingerprint::RequestFingerprint\` (now in openapi)
- \`crate::openapi::response::AiGenerator\` -> \`crate::response::AiGenerator\`
- \`crate::openapi::\{OpenApiOperation,OpenApiRoute,OpenApiSchema,OpenApiSpec\}\`
  -> \`crate::\{...\}\` (all local to openapi)
- \`crate::intelligent_behavior::{Request, Persona}\` ->
  \`mockforge_foundation::intelligent_behavior::\`
- \`crate::{Error, Result}\` -> \`mockforge_foundation::error::\`

mockforge-openapi Cargo.toml gains: \`once_cell\`, \`url\`, \`base64\`, and
an optional \`rayon\` dep gated behind a new \`parallel-routes\` feature
(previously a silent broken cfg on core's \`rayon\` feature). The cfg
gate in \`registry.rs\` is renamed \`feature = "rayon"\` ->
\`feature = "parallel-routes"\` accordingly.

\`mockforge_core::openapi_routes::\*\` now resolves via a re-export to
\`mockforge_openapi::openapi_routes\`, so every existing consumer import
path (\`OpenApiRouteRegistry\`, \`ValidationOptions\`, \`ValidationMode\`,
\`create_registry_from_file\`, \`create_registry_from_json\`, etc.)
continues to work without a single caller change.

With this, \`mockforge-core::openapi\` is a pure re-export facade. The
\`mockforge-openapi\` crate owns the complete OpenAPI domain model:
spec loading, schema validation, multi-spec merging, response
generation, route construction, router building, fixture matching,
request fingerprinting, and response tracing.

Verification:
- cargo build --workspace --exclude mockforge-desktop clean
- cargo test -p mockforge-openapi --lib — 160 passed (up from 106; 54
  tests moved in with openapi_routes)
- cargo test -p mockforge-core --lib — 1580 passed (down from 1634;
  the delta moved out with openapi_routes)
- cargo clippy --all-targets --all-features --workspace --exclude mockforge-desktop -- -D warnings clean
- cargo fmt --all --check clean

Plan: ~/.claude/plans/openapi-spec-extraction.md (Phase 4 complete)
